[frontend] refactor cards components (#4095)
