+++ opencti-platform/opencti-front/src/private/components/arsenal/malwares/MalwaresCards.jsx
@@ -3,7 +3,8 @@ import * as PropTypes from 'prop-types';
 import { graphql, createPaginationContainer } from 'react-relay';
 import { pathOr } from 'ramda';
 import ListCardsContent from '../../../../components/list_cards/ListCardsContent';
-import { MalwareCard, MalwareCardDummy } from './MalwareCard';
+import MalwareCard from './MalwareCard';
+import { GenericAttackCardDummy } from '../../common/cards/GenericAttackCard';
 import { setNumberOfElements } from '../../../../utils/Number';
 import { QueryRenderer } from '../../../../relay/environment';
 import StixDomainObjectBookmarks, {
@@ -57,11 +58,11 @@ class MalwaresCards extends Component {
                 this.props.data,
               )}
               CardComponent={<MalwareCard />}
-              DummyCardComponent={<MalwareCardDummy />}
+              DummyCardComponent={<GenericAttackCardDummy />}
               nbOfCardsToLoad={nbOfCardsToLoad}
               onLabelClick={onLabelClick.bind(this)}
               bookmarkList={bookmarks}
-              rowHeight={340}
+              rowHeight={350}
             />
           </div>
         )}
